scripts/requirements.txt
androguard==3.3.5
yara-python>=4.2.3
PyYAML==6.0
